(SERVER-887) Implement get-code-content
This commit implements the open source version of get-code-content. Like
code-id, it shells out to a script that is expected to put file content
on STDOUT. Unlike code-id, it returns an InputStream that can be used in
an HTTPResponse.

This commit updates ShellUtils to use an InputStream inside of
ExecutionResult.

Finally, the stream using in ShellUtils should ideally consist of piped
i/o streams, but that was considered out of scope given that it required
some threading work.