(SERVER-1131) Throw a better error message for commands with arguments
Previously if a command with arguments were passed to
shell-utils/execute-command, it would throw an error about
'/usr/bin/echo foo' command does not exist. This commit updates the
validation to throw an error if the command does not exist, but the
first part when split on a space does exist.