(SERVER-1571) Refactor ShellUtils, push args to ExecutionOptions
We were starting to have a nasty proliferation of method signatures
in the Java ShellUtils code, because there are several optional
arguments that can be passed in related to how to execute a process.

This commit pushes all of the arguments up into the ExecutionOptions
class, and reduces the number of method signatures to just include
variants with and without an instance of that class.