(SERVER-1671) Don't log stdout from shell utils
Previously, when ShellUtils was configured to combine both stderr and
stdout from the executed command, the combination of the two was
logged as a warn level message.  With the changes in this commit, none
of the info written to stdout is logged.  Only the info written to
stderr, if applicable, is logged.