(MAINT) Allow stderr/out in combined output test to be in either order
Previously, the `combines-stderr-and-stdout-correctly` test expected the
output order in which stderr and stdout was captured to match the order
in which the spawned executable wrote to each stream.  The order in
which the Java process receives content from the streams, however, is
not guaranteed to match the order in which the content was originally
written.  For this reason, the test was changed to allow the content
from the streams to arrive in either order.