(SERVER-2232) Fix unit test failure on OSX
Prior to this commit, the script used ruby to get the environment keys
to work around a problem discovered during SERVER-1494. By switching to
ruby we make unintended calls to CoreFoundations on OSX which adds an
additional environment variables __CF_USER_TEXT_ENCODING causing the
tests to fail only on OSX.

To fix this, this commit reverts the test script to the shell
and uses the cut tool to select the keys.